Refactor TreeView stories to align with ADR
